:hammer: Relocate a container generator utility function
